Renamed tasks folder to src
