+++ README.md
@@ -13,13 +13,13 @@ pip install -r requirements.txt
 ```
 4. Download a build of FFMPEG for your operating system on https://ffmpeg.org/download.html
 5. Extract the zip file (preferably to the mbox directory)
-6. Navigate to tasks/constants.py and change variable FFMPEG_PATH to the executable of where you extracted the zip file.
+6. Navigate to src/constants.py and change variable FFMPEG_PATH to the executable of where you extracted the zip file.
 7. Add a new application to your discord account on https://discord.com/developers/applications
 ![image](images/install1.png)
 The name you choose will be the bot's name. (recommended to name 'Music Box')
 ![image](images/install2.png)
 Click on your application and then (1.) click on Bot tab and (2.) click on copy button under token
-8. Navigate to tasks/constants.py and change variable TOKEN to the copied value.
+8. Navigate to src/constants.py and change variable TOKEN to the copied value.
 
 ## Usage
 Run main.py on the terminal with python
